Method search is added

+++ src/main/resources/schema.sql
@@ -13,9 +13,9 @@ create table flights
     foreign key (aircraft_id) references airCrafts,
     source_city      varchar(255) not null,
     destination_city varchar(255) not null,
-    departure_time   integer      not null,
-    arrival_time     integer      not null,
-    journey_duration integer      not null,
+    departure_time   BIGINT      not null,
+    arrival_time     BIGINT      not null,
+    journey_duration BIGINT      not null,
     price            integer      not null
 );
